WIP not compiling

+++ Platforms/ZXBox.Blazor/Hardware/Kempston.cs
@@ -11,7 +11,7 @@ namespace ZXBox.Hardware.Input.Joystick
     /// <summary>
     /// The Kempston joystick interface differs from the other common types in that it does not map to the ZX Spectrum keyboard directly. Rather, it maps to a particular hardware port (0x1f) and support must therefore be 'built-in' to the software. Fortunately, the Kempston joystick interface was enormously popular, and support was very easy to provide, making Kempston control a common, almost standard, feature of most games.
     /// Assuming an appropriate interface is attached, reading from port 0x1f returns the current state of the Kempston joystick in the form 000FUDLR, with active bits high.
-    /// © www.worldofspectrum.com
+    /// ï¿½ www.worldofspectrum.com
     /// </summary>
 
     public class Kempston:IInput
@@ -38,7 +38,7 @@ namespace ZXBox.Hardware.Input.Joystick
 
         #region IInput Members
 
-        public int Input(int Port,int tact)
+        public byte Input(int Port,int tact)
         {
             int returnvalue = 0xFF;
             if ((Port &0xff) == 0x1f)
@@ -48,7 +48,7 @@ namespace ZXBox.Hardware.Input.Joystick
                 this.Gamepad = Gamepads.FirstOrDefault();
                 if (Gamepad == null)
                 {
-                    return returnvalue;
+                    return (byte)returnvalue;
                 }
                 //GamePadState currentState = GamePad.GetState(playerIndex);
                 //000FUDLR
@@ -66,7 +66,7 @@ namespace ZXBox.Hardware.Input.Joystick
                         returnvalue |= 1;
                 }
             }
-            return returnvalue;
+            return (byte)returnvalue;
         }
 
         #endregion
