+++ ZXBox.Core/Cpus/Sharp/LR35902.cs
@@ -5,12 +5,12 @@ namespace ZXBox.Core.Cpus.Sharp
 {
     public class LR35902 : Z80
     {
-        public override int ReadByteFromMemory(int address)
+        public override byte ReadByteFromMemory(int address)
         {
             throw new System.NotImplementedException();
         }
 
-        public override void WriteByteToMemory(int address, int bytetowrite)
+        public override void WriteByteToMemory(int address, byte bytetowrite)
         {
             throw new System.NotImplementedException();
         }
@@ -64,14 +64,14 @@ namespace ZXBox.Core.Cpus.Sharp
                         switch (opcode)
                         {
                             case 0x08:
-                                WriteByteToMemory(GetNextPCWord(), SP);
+                                WriteByteToMemory(GetNextPCWord(), (byte)SP);
                                 SubtractNumberOfTStatesLeft(13);
                                 break;
                             case 0x10://STOP
                                 Halt();
                                 break;
                             case 0x22://LDI(HL),A
-                                WriteByteToMemory(HL, A);
+                                WriteByteToMemory(HL, (byte)A);
                                 HL++;
                                 break;
                             case 0x2A://LDI  A,(HL)
@@ -79,7 +79,7 @@ namespace ZXBox.Core.Cpus.Sharp
                                 HL++;
                                 break;
                             case 0x32://LDD(HL),A
-                                WriteByteToMemory(HL, A);
+                                WriteByteToMemory(HL, (byte)A);
                                 HL--;
                                 break;
                             case 0x3A://LDD  A,(HL)
