+++ ZXBox.Core/Hardware/Interfaces/IInput.cs
@@ -2,7 +2,7 @@ namespace ZXBox.Hardware.Interfaces;
 
 public interface IInput
 {
-    int Input(int Port, int tact);
+    byte Input(int Port, int tact);
     void AddTStates(int tstates)
     { }
 }
